Remove blank space between vertical image thumbnails in the sheets.
This should also have the side effect of making it work with images of a
larger aspect ratio (widescreen).  The whole concept of a static (or max)
thumbnail width is bogus.

This not only makes the sheets slightly smaller (despite the fact blank
space compresses well, but also makes the code simpler.  Dig the diffstat.